backend/requirements/base.txt
# Django Core
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# Authentication
djangorestframework-simplejwt==5.3.1

# Celery
celery==5.3.6
redis==5.0.1
django-redis==5.4.0

# Utilities
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
simpleeval==1.0.0
jinja2==3.1.3
requests==2.31.0
Pillow==10.2.0
qrcode==7.4.2

# Validation
django-filter==23.5
marshmallow==3.20.1

# Monitoring
django-prometheus==2.3.1

# DRF Extensions
drf-spectacular==0.27.0
djangorestframework-camel-case==1.4.2
